First Level Placeholder
I personally have thought and planned out a good idea for this, 

Its not really going out on a limb because of features but its going to be a physics puzzle. 
I'll be making the assets and hooked up with my friend who is a GREAT artist to supply some background work.